Safari-611.1.10/LayoutTests/platform/glib/http/tests/websocket/tests/hybi/stop-on-resume-in-error-handler-expected.txt

Tests that WebSocket correctly handles being stopped in the error event handler upon restoring from PageCache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
PASS Page did enter and was restored from the page cache
error event fired
PASS wasRestoredFromPageCache is true
PASS successfullyParsed is true

TEST COMPLETE
